UltraShip/ultra-doc-intelligence/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==0.8.11
huggingface_hub==0.20.3
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
groq==0.4.2 
python-dotenv==1.0.0
pydantic==2.4.2
chromadb==0.4.18

streamlit==1.28.1
requests==2.31.0
